+++ roman/errors_test.go
@@ -13,7 +13,7 @@ import (
 
 func Test_newNumberFormatError(t *testing.T) {
 	err := errors.New("parse error")
-	assert.Equal(t, &NumberFormatError{
+	assert.Equal(t, &NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "1h",
 		Err:   err,
@@ -22,7 +22,7 @@ func Test_newNumberFormatError(t *testing.T) {
 
 func Test_NumberFormatError_Unwrap(t *testing.T) {
 	err := errors.New("parse error")
-	assert.Equal(t, err, (&NumberFormatError{
+	assert.Equal(t, err, (&NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "1h",
 		Err:   err,
@@ -31,22 +31,22 @@ func Test_NumberFormatError_Unwrap(t *testing.T) {
 
 func Test_NumberFormatError_Error(t *testing.T) {
 	err := errors.New("parse error")
-	assert.Equal(t, `roman.UnmarshalJSON: invalid roman number`, (&NumberFormatError{
+	assert.Equal(t, `roman.UnmarshalJSON: invalid roman number`, (&NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "",
 		Err:   nil,
 	}).Error())
-	assert.Equal(t, `roman.UnmarshalJSON: parse error`, (&NumberFormatError{
+	assert.Equal(t, `roman.UnmarshalJSON: parse error`, (&NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "",
 		Err:   err,
 	}).Error())
-	assert.Equal(t, `roman.UnmarshalJSON: "1h": invalid roman number`, (&NumberFormatError{
+	assert.Equal(t, `roman.UnmarshalJSON: "1h": invalid roman number`, (&NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "1h",
 		Err:   nil,
 	}).Error())
-	assert.Equal(t, `roman.UnmarshalJSON: "1h": parse error`, (&NumberFormatError{
+	assert.Equal(t, `roman.UnmarshalJSON: "1h": parse error`, (&NumberFormatError[string]{
 		Func:  "UnmarshalJSON",
 		Input: "1h",
 		Err:   err,