+++ sem/compare_test.go
@@ -7,6 +7,8 @@ package sem
 import (
 	"testing"
 
+	"go.lstv.dev/util/constraint"
+
 	"github.com/stretchr/testify/assert"
 )
 
@@ -34,67 +36,67 @@ func Test_DefaultComparePreRelease(t *testing.T) {
 }
 
 func Test_CompareVersion(t *testing.T) {
-	testCompareVersion(t, CompareVersion)
+	testCompareVersion[string, string](t, CompareVersion[string, string])
 }
 
-func testCompareVersion(t *testing.T, f func(string, string) (int, error)) {
+func testCompareVersion[T1, T2 constraint.ParserInput](t *testing.T, f func(T1, T2) (int, error)) {
 	t.Helper()
-	c, err := f("0.0.0", "0.0.0")
+	c, err := f(T1("0.0.0"), T2("0.0.0"))
 	assert.Equal(t, 0, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0", "0.0.0")
+	c, err = f(T1("1.0.0"), T2("0.0.0"))
 	assert.Equal(t, 1, c)
 	assert.NoError(t, err)
-	c, err = f("0.0.0", "1.0.0")
+	c, err = f(T1("0.0.0"), T2("1.0.0"))
 	assert.Equal(t, -1, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0-alfa.1", "1.0.0-alfa.1")
+	c, err = f(T1("1.0.0-alfa.1"), T2("1.0.0-alfa.1"))
 	assert.Equal(t, 0, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0", "1.0.0-alfa.1")
+	c, err = f(T1("1.0.0"), T2("1.0.0-alfa.1"))
 	assert.Equal(t, 1, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0-alfa.1", "1.0.0")
+	c, err = f(T1("1.0.0-alfa.1"), T2("1.0.0"))
 	assert.Equal(t, -1, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0-alfa.2", "1.0.0-alfa.1")
+	c, err = f(T1("1.0.0-alfa.2"), T2("1.0.0-alfa.1"))
 	assert.Equal(t, 1, c)
 	assert.NoError(t, err)
-	c, err = f("1.0.0-alfa.1", "1.0.0-alfa.2")
+	c, err = f(T1("1.0.0-alfa.1"), T2("1.0.0-alfa.2"))
 	assert.Equal(t, -1, c)
 	assert.NoError(t, err)
-	c, err = f("1.0", "0.0.0")
+	c, err = f(T1("1.0"), T2("0.0.0"))
 	assert.Empty(t, c)
 	assert.Error(t, err)
-	c, err = f("0.0.0", "1.0")
+	c, err = f(T1("0.0.0"), T2("1.0"))
 	assert.Empty(t, c)
 	assert.Error(t, err)
 }
 
 func Test_CompareTag(t *testing.T) {
-	testCompareTag(t, CompareTag)
+	testCompareTag[string, string](t, CompareTag[string, string])
 }
 
-func testCompareTag(t *testing.T, f func(string, string) (int, error)) {
+func testCompareTag[T1, T2 constraint.ParserInput](t *testing.T, f func(T1, T2) (int, error)) {
 	t.Helper()
-	c, err := f("v0.0.0", "v0.0.0")
+	c, err := f(T1("v0.0.0"), T2("v0.0.0"))
 	assert.Equal(t, 0, c)
 	assert.NoError(t, err)
-	c, err = f("v1.0.0", "v0.0.0")
+	c, err = f(T1("v1.0.0"), T2("v0.0.0"))
 	assert.Equal(t, 1, c)
 	assert.NoError(t, err)
-	c, err = f("v0.0.0", "v1.0.0")
+	c, err = f(T1("v0.0.0"), T2("v1.0.0"))
 	assert.Equal(t, -1, c)
 	assert.NoError(t, err)
-	c, err = f("v1.0", "v0.0.0")
+	c, err = f(T1("v1.0"), T2("v0.0.0"))
 	assert.Empty(t, c)
 	assert.Error(t, err)
-	c, err = f("v0.0.0", "v1.0")
+	c, err = f(T1("v0.0.0"), T2("v1.0"))
 	assert.Empty(t, c)
 	assert.Error(t, err)
 }
 
 func Test_Compare(t *testing.T) {
-	testCompareVersion(t, Compare)
-	testCompareTag(t, Compare)
+	testCompareVersion[string, string](t, Compare[string, string])
+	testCompareTag[string, string](t, Compare[string, string])
 }