+++ sem/latest_test.go
@@ -7,16 +7,18 @@ package sem
 import (
 	"testing"
 
+	"go.lstv.dev/util/constraint"
+
 	"github.com/stretchr/testify/assert"
 )
 
 func Test_LatestVersion(t *testing.T) {
-	testLatestVersion(t, LatestVersion)
+	testLatestVersion[string, string](t, LatestVersion[string, string])
 }
 
-func testLatestVersion(t *testing.T, f func(string, string) (Ver, error)) {
+func testLatestVersion[T1, T2 constraint.ParserInput](t *testing.T, f func(T1, T2) (Ver, error)) {
 	t.Helper()
-	l, err := f("0.0.0", "0.0.0")
+	l, err := f(T1("0.0.0"), T2("0.0.0"))
 	assert.Equal(t, Ver{
 		Major:      0,
 		Minor:      0,
@@ -25,7 +27,7 @@ func testLatestVersion(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("1.0.0", "0.0.0")
+	l, err = f(T1("1.0.0"), T2("0.0.0"))
 	assert.Equal(t, Ver{
 		Major:      1,
 		Minor:      0,
@@ -34,7 +36,7 @@ func testLatestVersion(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("0.0.0", "1.0.0")
+	l, err = f(T1("0.0.0"), T2("1.0.0"))
 	assert.Equal(t, Ver{
 		Major:      1,
 		Minor:      0,
@@ -43,24 +45,24 @@ func testLatestVersion(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("1.0", "0.0.0")
+	l, err = f(T1("1.0"), T2("0.0.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
-	l, err = f("0.0.0", "1.0")
+	l, err = f(T1("0.0.0"), T2("1.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
-	l, err = f("0.0", "1.0")
+	l, err = f(T1("0.0"), T2("1.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
 }
 
 func Test_LatestTag(t *testing.T) {
-	testLatestTag(t, LatestTag)
+	testLatestTag[string, string](t, LatestTag[string, string])
 }
 
-func testLatestTag(t *testing.T, f func(string, string) (Ver, error)) {
+func testLatestTag[T1, T2 constraint.ParserInput](t *testing.T, f func(T1, T2) (Ver, error)) {
 	t.Helper()
-	l, err := f("v0.0.0", "v0.0.0")
+	l, err := f(T1("v0.0.0"), T2("v0.0.0"))
 	assert.Equal(t, Ver{
 		Major:      0,
 		Minor:      0,
@@ -69,7 +71,7 @@ func testLatestTag(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("v1.0.0", "v0.0.0")
+	l, err = f(T1("v1.0.0"), T2("v0.0.0"))
 	assert.Equal(t, Ver{
 		Major:      1,
 		Minor:      0,
@@ -78,7 +80,7 @@ func testLatestTag(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("v0.0.0", "v1.0.0")
+	l, err = f(T1("v0.0.0"), T2("v1.0.0"))
 	assert.Equal(t, Ver{
 		Major:      1,
 		Minor:      0,
@@ -87,18 +89,18 @@ func testLatestTag(t *testing.T, f func(string, string) (Ver, error)) {
 		Build:      "",
 	}, l)
 	assert.NoError(t, err)
-	l, err = f("v1.0", "v0.0.0")
+	l, err = f(T1("v1.0"), T2("v0.0.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
-	l, err = f("v0.0.0", "v1.0")
+	l, err = f(T1("v0.0.0"), T2("v1.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
-	l, err = f("v0.0", "v1.0")
+	l, err = f(T1("v0.0"), T2("v1.0"))
 	assert.Empty(t, l)
 	assert.Error(t, err)
 }
 
 func Test_Latest(t *testing.T) {
-	testLatestVersion(t, Latest)
-	testLatestTag(t, Latest)
+	testLatestVersion[string, string](t, Latest[string, string])
+	testLatestTag[string, string](t, Latest[string, string])
 }